+++ src/test/java/hibernate/ddl/CreateQueryBuilderTest.java
@@ -1,8 +1,9 @@
-package hibernate;
+package hibernate.ddl;
 
 import domain.Person;
 import domain.Person2;
 import domain.Person3;
+import hibernate.ddl.CreateQueryBuilder;
 import hibernate.entity.EntityClass;
 import org.junit.jupiter.api.Test;
