+++ src/test/java/hibernate/dml/DeleteQueryBuilderTest.java
@@ -1,5 +1,7 @@
-package hibernate;
+package hibernate.dml;
 
+import hibernate.entity.EntityObject;
+import hibernate.ddl.DeleteQueryBuilder;
 import jakarta.persistence.*;
 import org.junit.jupiter.api.Test;
